+++ papas/detectors/CMS.py
