+++ papas/detectors/base.py
