+++ papas/detectors/detector.py
