+++ papas/detectors/geometry.py
