+++ papas/detectors/material.py
