+++ papas/detectors/perfect.py
