+++ papas/detectors/test_geometry.py
